Moves useCells to Map
