+++ src/models/Logic/ILogic.ts
@@ -1,5 +1,5 @@
 import Direction from '../Direction';
-import {Action} from '../Actions';
+import { Action } from '../Actions';
 import LogicState from './LogicState';
 
 interface ILogic {
